requirements.txt
Flask==3.0.0
Flask-MySQLdb==1.0.1
pymysql==1.1.0
numpy==1.26.4
pandas==2.2.1
sentence-transformers==2.2.2
torch==2.1.2
transformers==4.36.2
scikit-learn==1.3.0
nltk==3.8.1
gunicorn==21.2.0
